Service Component Architecture
SCA is a programming model for abstracting business functions as
components and using them as building blocks to assemble business
solutions. An SCA component offers services and depends on functions
that are called references. It also has an implementation associated it
with it which is the business logic that can be implemented in any
technology.